✨ feat:  게임페이지 및 소켓(50%) 구현 🗑️ cleanup : 파일구조 조정
