+++ app/dashboard/game/layout.tsx
@@ -1,3 +1,4 @@
+import { SocketProvider } from "@/app/context/context";
 import NavBar from "../../components/NavBar";
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
@@ -5,7 +6,9 @@ export default function DashboardLayout({ children }: { children: React.ReactNod
     <div className="flex h-screen">
       <NavBar />
       <div className="flex-1 ml-[85px] xl:ml-64 min-h-screen overflow-y-auto">
-        {children}
+        <SocketProvider>
+          {children}
+        </SocketProvider>
       </div>
     </div>
   );
