+++ app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginCompo from "../components/login";
+import LoginCompo from "../components/Login";
 
 
 export default function LoginPage() {
