+++ app/register/page.tsx
@@ -1,4 +1,4 @@
-import Register from "../components/register";
+import Register from "../components/Register";
 
 
 export default function registerPage() {
